+++ src/lib/util/jsonconfig.cpp
@@ -58,7 +58,7 @@ string JSONConfig::getFullConfigurationString()
 
 		sort(parts.begin(), parts.end(), JSONEntryCompare);
 
-		for (int i = 0 ; i < parts.size() ; i++)
+		for (size_t i = 0 ; i < parts.size() ; i++)
 		{
 			result += parts[i];
 			if (i+1 != parts.size())
